+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2024.02.13',
+  '2024.02.16',
   // Fallback runtime versions.
+  '2024.02.13',
   '2024.02.09',
-  '2024.02.05',
 ];
 
 /// Sets the current runtime versions.
@@ -63,9 +63,9 @@ bool shouldGCVersion(String version) =>
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.2.0';
 final String toolStableDartSdkVersion = '3.3.0';
-final String toolStableFlutterSdkVersion = '3.16.9';
+final String toolStableFlutterSdkVersion = '3.19.0';
 final String toolPreviewDartSdkVersion = '3.4.0-131.0.dev';
-final String toolPreviewFlutterSdkVersion = '3.19.0-0.4.pre';
+final String toolPreviewFlutterSdkVersion = '3.20.0-1.1.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);